+++ Sora/VideoCapturer.swift
@@ -3,22 +3,22 @@ import WebRTC
 
 /**
  映像キャプチャーのイベントハンドラです。
+ 廃止されました。
  */
-public final class VideoCapturerHandlers {
-    
-    /// このプロパティは onCapture に置き換えられました。
-    @available(*, deprecated, renamed: "onCapture",
-    message: "このプロパティは onCapture に置き換えられました。")
-    public var onCaptureHandler: ((VideoFrame) -> Void)? {
-        get { onCapture }
-        set { onCapture = newValue }
-    }
-    
-    /// 映像フレームの生成時に呼ばれるクロージャー
-    public var onCapture: ((VideoFrame) -> Void)?
+@available(*, unavailable, message: "VideoCapturerHandlers は廃止されました。 CameraVideoCapturerHandlers を利用してください。")
+public final class VideoCapturerHandlers {}
 
-    /// 初期化します。
-    public init() {}
+public class CameraVideoCapturerHandlers {
+    
+    // 生成された映像フレームを受け取る
+    // 返した映像フレームがストリームに渡される
+    public var onCapture: ((VideoFrame) -> VideoFrame)?
+
+    // カメラの起動時に呼ばれる
+    public var onStart: (() -> Void)?
+    
+    // カメラの停止時に呼ばれる
+    public var onStop: (() -> Void)?
     
 }
 
@@ -30,21 +30,11 @@ public final class VideoCapturerHandlers {
  映像キャプチャーとデータの送受信は別であることに注意してください。
  映像キャプチャーが映像フレームを生成しても、
  メディアストリームなしではサーバーに映像が送信されません。
+ 
+ 廃止されました。
  */
-public protocol VideoCapturer: class {
-    
-    /// 映像フレームを渡すストリーム
-    var stream: MediaStream? { get set }
-    
-    /// イベントハンドラ
-    var handlers: VideoCapturerHandlers { get }
-    
-    /// 映像キャプチャーを起動します。
-    func start()
-    
-    /// 映像キャプチャーを停止します。
-    func stop()
-    
+@available(*, unavailable, message: "VideoCapturer は廃止されました。")
+public protocol VideoCapturer: AnyObject {
 }
 
 /**
@@ -52,7 +42,7 @@ public protocol VideoCapturer: class {
  `MediaStream.videoFilter` にセットすると、
  生成された映像フレームはこのプロトコルの実装によって加工されます。
  */
-public protocol VideoFilter: class {
+public protocol VideoFilter: AnyObject {
     
     /**
      映像フレームを加工します。